requirements.txt
celery==5.3.6
django==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.15.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9